requirements.txt
streamlit==1.52.2
pandas==2.3.3
numpy==2.4.0
scikit-learn==1.8.0
joblib==1.4.2
pyyaml==6.0.3
openpyxl==3.1.5
fpdf2==2.8.1
altair==5.3.0
